requirement-prod.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-core==2.14.0
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
cachetools==5.3.1
python-dateutil==2.8.2
email-validator==2.1.0
phonenumbers==8.13.22
regex==2023.10.3
gunicorn==21.2.0